% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{theme_weather}
\alias{theme_weather}
\title{theme_weather}
\usage{
theme_weather(base_size = 9)
}
\arguments{
\item{base_size}{base font size (default is 9)}
}
\description{
A custom ggplot2 theme used throughout this project
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
  geom_point() +
  theme_weather()
}
